+++ test/src/FilterTest.h
@@ -5,7 +5,7 @@
 
 #include "cinder/gl/TextureFont.h"
 
-#include "../../lib/Cinder-FileWatcher/src/mason/FileWatcher.h" // TEMP
+#include "../../blocks/Cinder-FileWatcher/src/mason/FileWatcher.h" // TEMP
 
 class FilterSinglePass : public ui::Filter {
 public:
@@ -19,51 +19,6 @@ public:
 private:
 };
 
-typedef std::shared_ptr<class FilterBlur>	FilterBlurRef;
-
-class FilterBlur : public ui::Filter {
-  public:
-	FilterBlur();
-
-	void configure( const ci::ivec2 &size, ui::Filter::PassInfo *info ) override;
-	void process( ui::Renderer *ren, const ui::Filter::Pass &frame  ) override;
-
-	ci::gl::GlslProgRef	mGlsl;
-	mason::ScopedWatch mWatchGlsl;
-
-	const ci::vec2&	getBlurPixels() const						{ return mBlurPixels; }
-	void			setBlurPixels( const ci::vec2 &pixels )		{ mBlurPixels = pixels; }
-
-  private:
-	ci::vec2	mBlurPixels = ci::vec2( 1 );
-};
-
-typedef std::shared_ptr<class FilterDropShadow>	FilterDropShadowRef;
-
-class FilterDropShadow : public ui::Filter {
-public:
-	FilterDropShadow();
-
-	void configure( const ci::ivec2 &size, ui::Filter::PassInfo *info ) override;
-	void process( ui::Renderer *ren, const ui::Filter::Pass &frame  ) override;
-
-	ci::gl::GlslProgRef	mGlsl;
-	mason::ScopedWatch mWatchGlsl;
-
-	void			setBlurPixels( const ci::vec2 &pixels )		{ mBlurPixels = pixels; }
-	const ci::vec2&	getBlurPixels() const { return mBlurPixels; }
-
-	void			setShadowOffset( const ci::vec2 &pixels )	{ mShadowOffset= pixels; }
-	const ci::vec2&	getShadowOffset() const { return mShadowOffset; }
-
-	void			setDownsampleFactor( float factor );
-	float			getDownsampleFactor() const			{ return mDownsampleFactor; }
-
-private:
-	ci::vec2	mBlurPixels = ci::vec2( 1 );
-	ci::vec2	mShadowOffset = ci::vec2( 10 );
-	float		mDownsampleFactor = 1;
-};
 
 class FilterTest : public ui::SuiteView {
   public:
@@ -73,10 +28,14 @@ class FilterTest : public ui::SuiteView {
 
   private:
 	bool keyDown( ci::app::KeyEvent &event ) override;
+	void loadGlsl();
 
 	std::shared_ptr<FilterSinglePass>	mFilterSinglePass;
-	FilterBlurRef						mFilterBlur;
-	FilterDropShadowRef					mFilterDropShadow;
+	ui::FilterBlurRef					mFilterBlur;
+	ui::FilterDropShadowRef				mFilterDropShadow;
+
+	ci::gl::GlslProgRef					mGlslBlur, mGlslDropshadow;
+	mason::ScopedWatch					mWatchGlslBlur, mWatchGlslDropshadow;
 
 	ui::ViewRef				mContainerView;
 	ui::ImageViewRef        mImageView;
